requirements.txt
appdirs==1.4.3
click==6.7
decorator==4.0.11
Django==1.11
djangorestframework==3.6.3
geocoder==1.22.4
packaging==16.8
pkg-resources==0.0.0
pyparsing==2.2.0
pytz==2017.2
ratelim==0.1.6
requests==2.14.2
six==1.10.0
